Presentation



See merge request !100

+++ gatorslist/application/view/useraccount/index.php
@@ -1,4 +1,4 @@
-<div class="container" style="margin-bottom: 50px;">
+<div class="container" id="account-overview" style="margin-bottom: 50px;">
     <div class="box">
         <div class="col-lg-8 col-lg-offset-2" style="margin-top:1%;" >
             <div class="panel-heading" style="text-align:center;color:#ffdf80;"><h2>Account overview</h2></div>
@@ -43,7 +43,7 @@
 
                                 <td><a href="<?php echo URL . 'useraccount/editItem/' . htmlspecialchars($userproduct->Product_id, ENT_QUOTES, 'UTF-8'); ?>">EDIT</a></td>
  
-                                 <td><a href="<?php echo URL . 'useraccount/deleteItem/' . htmlspecialchars($userproduct->Product_id, ENT_QUOTES, 'UTF-8'); ?>">DELETE</a></td>
+                                 <td><a onclick="return confirm('This will permanantly delete your item.')" href="<?php echo URL . 'useraccount/deleteItem/' . htmlspecialchars($userproduct->Product_id, ENT_QUOTES, 'UTF-8'); ?>">DELETE</a></td>
                                 
                             </tr>
                         <?php } ?>
